requirements.txt
Flask-RESTful==0.3.9
uWSGI==2.0.20
Flask-HTTPAuth==4.7.0
